refactor(testes): exclui testes duplicados passando a ter 75 testes e mantendo a cobertura de 87,5%
